fix variable conversion error

+++ Snake.cpp
@@ -105,8 +105,9 @@ void Snake::moveX(int dx) {
 const std::vector<SDL_Rect> &Snake::getBody() const { return body; }
 
 void Snake::increaseSize() {
-  body.push_back({body[body.size() - 1].x - BASIC_UNITY_SIZE * body.size(),
-                  body[0].y, BASIC_UNITY_SIZE, BASIC_UNITY_SIZE});
+  body.push_back(
+      {(int)body[body.size() - 1].x - BASIC_UNITY_SIZE * (int)body.size(),
+       (int)body[0].y, BASIC_UNITY_SIZE, BASIC_UNITY_SIZE});
 }
 
 void Snake::updateBodyPositions() {}
